added image component to ../components/activities
